+++ Robust.Client/UserInterface/Controls/SplitContainer.cs
@@ -67,8 +67,11 @@ namespace Robust.Client.UserInterface.Controls
             get => _splitStart + _splitWidth / 2;
             set
             {
+                var args = new SplitCenterChangingEventArgs(value, _dragging);
+                OnSplitCenterChanging?.Invoke(args);
+
                 State = SplitState.Manual;
-                _splitStart = value - _splitWidth / 2;
+                _splitStart = args.SplitCenter - _splitWidth / 2;
                 ClampSplitCenter();
                 InvalidateMeasure();
                 OnSplitResized?.Invoke();
@@ -137,6 +140,7 @@ namespace Robust.Client.UserInterface.Controls
 
         public (Control First, Control Second)? Splits => ChildCount < 3 ? null : (GetChild(0), GetChild(1));
 
+        public event Action<SplitCenterChangingEventArgs>? OnSplitCenterChanging;
         public event Action? OnSplitResizeFinished;
         public event Action? OnSplitResized;
 
@@ -459,6 +463,18 @@ namespace Robust.Client.UserInterface.Controls
         {
             Horizontal,
             Vertical
+        }
+
+        public sealed class SplitCenterChangingEventArgs
+        {
+            public SplitCenterChangingEventArgs(float splitCenter, bool dragging)
+            {
+                SplitCenter = splitCenter;
+                Dragging = dragging;
+            }
+
+            public float SplitCenter { get; set; }
+            public bool Dragging { get; }
         }
 
         /// <summary>